requirements.txt
# Core
langgraph>=0.2.0
langchain-google-genai>=2.0.0
langchain-core>=0.3.0
google-generativeai>=0.8.0

# Vector DB
chromadb>=0.5.0

# UI
streamlit>=1.40.0

# Config
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.24.0
pytest-cov>=5.0.0

# Utilities
httpx>=0.27.0
tenacity>=9.0.0
